// packages/twenty-apps/internal/twenty-for-twenty/src/modules/resend/types/update-template.dto.ts
import type { EmailsField } from 'src/modules/resend/types/emails-field';

export type UpdateTemplateDto = {
  name: string;
  alias: string;
  status: string;
  fromAddress: EmailsField;
  subject: string;
  replyTo: EmailsField;
  htmlBody: string;
  textBody: string;
  resendUpdatedAt: string;
  publishedAt: string | null;
};